Expand details

+++ src/components/animal/AnimalDetail.js
@@ -4,21 +4,27 @@ import "./AnimalDetail.css";
 import { firstLetterCase } from '../../modules/helpers';
 
 const AnimalDetail = props => {
-    const [animal, setAnimal] = useState({ name: "", breed: "" });
+    const [animal, setAnimal] = useState({ name: "", breed: "", gender: "", age: "", weight: "", petOwner: "" });
     const [isLoading, setIsLoading] = useState(true);
 
     const handleDelete = () => {
-        setIsLoading(true);
-        AnimalManager.delete(props.animalId).then(() =>
-            props.history.push("/animals")
-        );
+        if(window.confirm("Are you sure you want to discharge this animal?")) {
+            setIsLoading(true);
+            AnimalManager.delete(props.animalId).then(() =>
+                props.history.push("/animals")
+            );
+        }
     };
 
     useEffect(() => {
         AnimalManager.get(props.animalId).then(animal => {
             setAnimal({
                 name: animal.name,
-                breed: animal.breed
+                breed: animal.breed,
+                gender: animal.gender,
+                age: animal.age,
+                weight: animal.weight,
+                petOwner: animal.petOwner
             });
             setIsLoading(false);
         });
@@ -32,20 +38,35 @@ const AnimalDetail = props => {
         )
     } else {
         return (
-            <div className="card">
-                <div className="card-content">
-                    <picture>
-                        <img src={require("./dog.svg")} alt="My Dog" />
-                    </picture>
-                    <h3>
-                        Name: <span style={{ color: "darkslategrey" }}>{firstLetterCase(animal.name)}</span>
-                    </h3>
-                    <p>Breed: {firstLetterCase(animal.breed)}</p>
-                    <button type="button" disabled={isLoading} onClick={handleDelete}>
-                        Discharge
-            </button>
+            <>
+                <button type="button"
+                    className="back"
+                    onClick={() => { props.history.push("/animals") }}>
+                    Go Back
+                </button>
+                <div className="card">
+                    <div className="card-content">
+                        <picture>
+                            <img src={require("./dog.svg")} alt="My Dog" />
+                        </picture>
+                        <h3>
+                            Name: <span style={{ color: "darkslategrey" }}>{firstLetterCase(animal.name)}</span>
+                        </h3>
+                        <p>Breed: {firstLetterCase(animal.breed)}</p>
+                        <p>Gender: {firstLetterCase(animal.gender)}</p>
+                        <p>Age: {animal.age}</p>
+                        <p>Weight: {animal.weight} pounds</p>
+                        <p>Owner: {firstLetterCase(animal.petOwner)}</p>
+                        <button type="button"
+                            onClick={() => props.history.push(`/animals/${props.animalId}/edit`)}>
+                            Edit
+                    </button>
+                        <button type="button" disabled={isLoading} onClick={handleDelete}>
+                            Discharge
+                    </button>
+                    </div>
                 </div>
-            </div>
+            </>
         );
     }
 };
