+++ src/components/location/LocationDetail.js
@@ -3,14 +3,16 @@ import LocationManager from '../../modules/LocationManager';
 import { firstLetterCase } from '../../modules/helpers';
 
 const LocationDetail = props => {
-    const [location, setLocation] = useState({ name: "" });
+    const [location, setLocation] = useState({ name: "", phoneNumber: "", address: "", hours: "" });
     const [isLoading, setIsLoading] = useState(true);
 
     const handleDelete = () => {
-        setIsLoading(true);
-        LocationManager.delete(props.locationId).then(() =>
-            props.history.push("/locations")
-        );
+        if (window.confirm("Are you sure you want to close this location?")) {
+            setIsLoading(true);
+            LocationManager.delete(props.locationId).then(() =>
+                props.history.push("/locations")
+            );
+        }
     };
 
     useEffect(() => {
@@ -18,6 +20,9 @@ const LocationDetail = props => {
             .then(location => {
                 setLocation({
                     name: location.name,
+                    phoneNumber: location.phoneNumber,
+                    address: location.address,
+                    hours: location.hours
                 });
                 setIsLoading(false);
             });
@@ -31,17 +36,31 @@ const LocationDetail = props => {
         )
     } else {
         return (
-            <div className="card">
-                <div className="card-content">
-                    <picture>
-                        <img src={require("./nashville-skyline.jpg")} alt="Nashville Skyline" />
-                    </picture>
-                    <h3>Location: <span style={{ color: 'darkslategrey' }}>{firstLetterCase(location.name)}</span></h3>
-                    <button type="button" disabled={isLoading} onClick={handleDelete}>
-                        Close
+            <>
+                <button type="button"
+                    className="back"
+                    onClick={() => { props.history.push("/locations") }}>
+                    Go Back
+                </button>
+                <div className="card">
+                    <div className="card-content">
+                        <picture>
+                            <img src={require("./nashville-skyline.jpg")} alt="Nashville Skyline" />
+                        </picture>
+                        <h3>Location: <span style={{ color: 'darkslategrey' }}>{firstLetterCase(location.name)}</span></h3>
+                        <p>Phone Number: {location.phoneNumber}</p>
+                        <p>Address: {location.address}</p>
+                        <p>Hours: {location.hours}</p>
+                        <button type="button"
+                            onClick={() => props.history.push(`/locations/${props.locationId}/edit`)}>
+                            Edit
                     </button>
+                        <button type="button" disabled={isLoading} onClick={handleDelete}>
+                            Close
+                    </button>
+                    </div>
                 </div>
-            </div>
+            </>
         );
     }
 }
